+++ pyron-engine/src/main/scala/com/cloudentity/pyron/openapi/OpenApiServiceVerticle.scala
@@ -2,7 +2,6 @@ package com.cloudentity.pyron.openapi
 
 import com.cloudentity.pyron.apigroup.{ApiGroup, ApiGroupConf, ApiGroupsChangeListener, ApiGroupsStore}
 import com.cloudentity.pyron.client.TargetClient
-import com.cloudentity.pyron.config.Conf
 import com.cloudentity.pyron.domain.flow._
 import com.cloudentity.pyron.domain.http.{RelativeUri, TargetRequest}
 import com.cloudentity.pyron.domain.openapi._
@@ -127,17 +126,17 @@ object OpenApiRuleBuilder {
 
     serviceIdOpt.map(serviceId =>
       OpenApiRule(
-        method         = r.rule.criteria.method,
-        serviceId      = serviceId,
-        group          = groupMatchCriteria,
-        pathPrefix     = PathPrefix(r.rule.criteria.rewrite.pathPrefix),
-        pathPattern    = PathPattern(r.rule.criteria.rewrite.matchPattern),
+        method = r.rule.criteria.method,
+        serviceId = serviceId,
+        group = groupMatchCriteria,
+        pathPrefix = PathPrefix(r.rule.criteria.rewrite.pathPrefix),
+        pathPattern = PathPattern(r.rule.criteria.rewrite.originPattern),
         dropPathPrefix = r.rule.dropPathPrefix,
-        rewritePath    = r.rule.rewritePath,
-        rewriteMethod  = r.rule.rewriteMethod,
-        plugins        = r.requestPlugins.toList ::: r.responsePlugins.toList,
-        tags           = r.rule.tags,
-        operationId    = r.rule.ext.openapi.flatMap(_.operationId)
+        rewritePath = r.rule.rewritePath,
+        rewriteMethod = r.rule.rewriteMethod,
+        plugins = r.requestPlugins.toList ::: r.responsePlugins.toList,
+        tags = r.rule.tags,
+        operationId = r.rule.ext.openapi.flatMap(_.operationId)
       )
     )
   }